backend/requirements.txt
# requirements.txt - Clean, compatible versions
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
mongoengine==0.27.0
djongo==1.3.6
pymongo==4.5.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
websockets==12.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
yfinance==0.2.33
python-decouple==3.8
daphne==4.0.0
asgiref==3.7.2